angler: Set up vendor partition image generation
This commit follows up the device tree change to include the vendor
partition in builds. Additionally, the IMS files are moved over to
the vendor partition entirely to see if that works fine, since those
files are not in /system on the factory image.

All blobs are sourced from the 6.0.1 (MTC19V) image.

Change-Id: I752b5325e2e20e0ff1c5af3cfc2dd2834d878085